Minor bugfixes and improvements
- Updated CarModelConverter fallback to include car id
- Updated PedRemapConverter to cast to uint8_t, not uint32_t
- Added WeaponConverter -1 case
- Added missing returns in ModMenuModule::PlayerAppearanceCheat
- Added logging in ModMenuModule::SavedCarsCheat
- Fixed rare ModMenuModule::MenuManager crash on exit
- Added missing conditional in ModMenuModule::PlayerPosMenu

+++ src/converters/ped-remap.h
@@ -5,7 +5,7 @@
 class PedRemapConverter {
 public:
 	static std::wstring ConvertToString(Game::PED_REMAP value) {
-		int32_t id = static_cast<int32_t>(value);
+		int8_t id = static_cast<int8_t>(value);
 
 		switch (id) {
 			case -1: return L"Default";
